refactor : 로또 구매할 때 자동 생성되는 티켓의 타입을 LottoBoughtTicket으로 변경(LottoTicket을 상속받음)
