+++ src/main/java/lotto/domain/LottoTicket.java
@@ -4,7 +4,6 @@ import java.util.*;
 
 public class LottoTicket {
 
-    private static final int TOTAL_NUMBER_COUNT = 12;
     private static final int VALID_NUMBER_COUNT = 6;
     private static final String NULL_ERROR_MESSAGE = "null값은 허용하지 않습니다.";
     private static final String EMPTY_ERROR_MESSAGE = "숫자는 하나 이상이어야 합니다.";
@@ -19,16 +18,8 @@ public class LottoTicket {
         this.lottoNumbers = lottoNumbers;
     }
 
-    public LottoRank getRank(final LottoWinner lottoWinner) {
-        int matchCount = calculateMatchCount(lottoWinner.getLottoWinnerTicket());
-        boolean matchBonusNumber = lottoNumbers.contains(lottoWinner.getLottoWinnerBonusNumber());
-        return LottoRank.matchLottoRank(matchCount, matchBonusNumber);
-    }
-
-    private int calculateMatchCount(final LottoWinnerTicket lottoWinnerTicket) {
-        Set<LottoNumber> matchingCheckContainer = new HashSet<>(lottoNumbers);
-        matchingCheckContainer.addAll(lottoWinnerTicket.getLottoNumbers());
-        return TOTAL_NUMBER_COUNT - matchingCheckContainer.size();
+    public boolean isContain(final LottoNumber lottoNumber) {
+        return this.lottoNumbers.contains(lottoNumber);
     }
 
     private void validateLottoTicket(final List<LottoNumber> lottoNumbers) {
